Adding terrain

+++ GameEngine/src/entities/Entity.java
@@ -9,10 +9,10 @@ public class Entity {
 	private TexturedModel model;
 	private Vector3f position;
 	private float rotX, rotY, rotZ;
-	private int scale;
+	private float scale;
 
-	public Entity(TexturedModel model, Vector3f position, float rotX, float rotY, float rotZ, int scale) {
-
+	public Entity(TexturedModel model, Vector3f position, float rotX, float rotY, float rotZ,
+			float scale) {
 		this.model = model;
 		this.position = position;
 		this.rotX = rotX;
@@ -22,19 +22,15 @@ public class Entity {
 	}
 
 	public void increasePosition(float dx, float dy, float dz) {
-
 		this.position.x += dx;
 		this.position.y += dy;
 		this.position.z += dz;
-
 	}
 
 	public void increaseRotation(float dx, float dy, float dz) {
-		
 		this.rotX += dx;
 		this.rotY += dy;
 		this.rotZ += dz;
-		
 	}
 
 	public TexturedModel getModel() {
@@ -77,12 +73,13 @@ public class Entity {
 		this.rotZ = rotZ;
 	}
 
-	public int getScale() {
+	public float getScale() {
 		return scale;
 	}
 
-	public void setScale(int scale) {
+	public void setScale(float scale) {
 		this.scale = scale;
 	}
 
 }
+
